+++ CHANGES.rst
@@ -26,6 +26,8 @@ Bugs Fixed
   similar behavior of the old publisher.
   [davisagli]
 
+- Use unicode transaction-notes to support ZODB 5.
+  [pbauer]
 
 Features Added
 ++++++++++++++